requirements.txt
streamlit==1.56.0
scikit-learn==1.8.0
pandas==3.0.2
joblib==1.5.3
fpdf2==2.8.7
